# docs/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: " Backup Test Restoration System"
host: localhost:4000
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /restores:
    get:
      description: Get all recent restore results
      summary: Get Restores
      operationId: GetRestores
      produces:
        - application/json
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  type: string
                timestamp:
                  type: integer
                result:
                  type: string
            example:
              - id: "1"
                timestamp: 1520114032
                result: SUCCESS 
              - id: "2"
                timestamp: 1520114255
                result: FAILURE
    post:
      description: Run a backup test restoration
      summary: Run Restore
      operationId: RunRestore
      consumes: application/json
      parameters:
        - in: body
          name: Restore
          description: The test restoration to run
          schema:
            type: object
            required:
              - location
              - file
              - dataType
              - decryptKey
            properties:
              location:
                type: string
              file:
                type: string
              dataType:
                type: string
              decryptKey:
                type: string
            example:
              location: "192.168.1.99"
              file: "/backup/backup.json.gpg"
              dataType: "json"
              decryptKey: "backupKey1"
      responses:
        '200':
           description: The restoration was successfully triggered
        '405':
          description: Invalid input
...